services/auth/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
PyJWT==2.8.0
python-jose
